+++ pyproject.toml
@@ -17,21 +17,14 @@ classifiers = [
 ]
 requires-python = ">=3.11"
 dependencies = [
-    "aiohttp>=3.11.18",
     "click>=8.1.8",
-    "db-dtypes>=1.4.3",
-    "greenlet>=3.2.2",
     "httpx>=0.28.1",
-    "jellyfish>=1.2.0",
     "mcp[cli]>=1.9.0",
     "numpy>=1.24.0",
     "pandas>=2.2.3",
-    "pyarrow<19.0.0",
     "pydantic>=2.11.4",
     "pytest>=8.3.5",
     "python-decouple>=3.8",
-    "scikit-learn>=1.6.1",
-    "rank-bm25>=0.2.2",
     "openapi-spec-validator>=0.7.2",
     "jsonref>=1.1.0",
     "pydantic-ai>=0.4.2",
@@ -41,6 +34,7 @@ dependencies = [
     "tabulate>=0.9.0",
     "ibis-framework>=10.6.0",
     "markitdown>=0.1.2",
+    "pyarrow<19.0.0",
 ]
 
 [project.scripts]
