+++ src/toolfront/__init__.py
@@ -1,5 +1,5 @@
-from toolfront.models.datasources.api import API
-from toolfront.models.datasources.database import Database
-from toolfront.models.datasources.library import Library
+from toolfront.models.api import API
+from toolfront.models.database import Database
+from toolfront.models.library import Library
 
 __all__ = ["API", "Database", "Library"]
