+++ src/toolfront/config.py
@@ -1,26 +1,11 @@
-# Number of items to return from search tools
-NUM_ENDPOINT_SEARCH_ITEMS = 100
-NUM_TABLE_SEARCH_ITEMS = 100
-NUM_DOCUMENT_SEARCH_ITEMS = 100
-NUM_DOCUMENT_SEARCH_ITEMS = 100
-NUM_QUERY_SEARCH_ITEMS = 25
-
-# TTL for spec downloads (24 hours)
-SPEC_DOWNLOAD_TTL = 24 * 60 * 60  # 24 hours in seconds
-
-# TTL for spec downloads (24 hours)
-SPEC_DOWNLOAD_TTL = 24 * 60 * 60  # 24 hours in seconds
-
-# Data
 MAX_DATA_ROWS = 100
+
 MAX_DATA_CHARS = 30000
 
-# ALRU Cache
-CACHE_TTL = 3600
-
-# Timeout
 TIMEOUT_SECONDS = 30.0
 
 DEFAULT_MODEL = "openai:gpt-4o"
 
 MAX_RETRIES = 3
+
+CHUNK_SIZE = 10000
