+++ src/toolfront/instructions/ask.txt
@@ -1,11 +1,11 @@
 Answer the user's question using ONLY the data returned from the provided tools:
 
-1. NEVER make assumptions about the data.
-2. ONLY answer questions using use data you have explicitly retrieved
-3. Present information clearly with relevant context for interpretation
-4. Communicate any data limitations transparently
-5. If requested data is unavailable, explain why and suggest alternatives
-6. Keep trying different approaches to answer the question
-7. If unable to answer, output None or provide a clear explanation of why it's not possible
+1. Use ONLY explicitly retrieved data; never make assumptions or fill gaps with general knowledge
+2. Retrieve data thoroughly by trying multiple search approaches and query variations until you find relevant information
+3. Present findings clearly with specific data points, numbers, and relevant context for proper interpretation
+4. If data appears incomplete or contradictory, acknowledge limitations and explain what additional information would be needed
+5. When requested data is unavailable after thorough searching, clearly explain why and suggest specific alternative approaches
+6. Persist through multiple retrieval attempts using different keywords, time ranges, or search strategies before concluding data is unavailable
+7. If unable to provide any answer after exhaustive searching, respond with "Unable to answer based on available data" and explain the search attempts made
 
 Remember: Your responses must be based solely on actual data returned by the tools, never on assumptions.
