--- src/toolfront/models/actions/__init__.py
@@ -1,15 +1,0 @@
-# from toolfront.tools.api import inspect_endpoint, request_api, search_endpoints
-# from toolfront.tools.database import inspect_table, query_database, sample_table, search_tables
-# from toolfront.tools.storage import inspect_document, read_storage, sample_document, search_documents
-
-# __all__ = [
-#     "database_tools",
-#     "api_tools",
-#     "storage_tools",
-# ]
-
-# database_tools = [inspect_table, query_database, sample_table, search_tables]
-
-# api_tools = [inspect_endpoint, request_api, search_endpoints]
-
-# storage_tools = [inspect_document, sample_document, search_documents, read_storage]
