--- src/toolfront/models/atomics/document.py
@@ -1,12 +1,0 @@
-from abc import ABC
-
-from pydantic import BaseModel, Field
-
-
-class Document(BaseModel, ABC):
-    """Abstract base class for documents."""
-
-    path: str = Field(
-        ...,
-        description="Document path in relative to the library url e.g. 'path/to/file.pdf'.",
-    )
